Move RoutesScreen to the routes:ui module.
