+++ android/routes/ui/src/main/java/com/trafi/routes/ui/internal/LocationDisplayText.kt
@@ -1,0 +1,10 @@
+package com.trafi.routes.ui.internal
+
+import com.trafi.core.android.model.Location
+
+internal val Location?.displayText: String
+    get() = this?.run {
+        name
+            ?: address
+            ?: coordinate.run { "$lat,$lng" }
+    } ?: ""
